[LETS-33] re-arrange sequential recovery redo code in preparation for the addition of parallel recovery redo infrastructure (#2616)
http://jira.cubrid.org/browse/LETS-33

This is a preparatory change to prepare for the addition of parallel recovery redo infrastructure.
Functionally, there is no change, mostly code reorganization.

log_recovery_redo.hpp:
- renamed header to `log_recovery_redo.hpp' because it contains all code related to [sequential/synchronous] recovery redo
- added a small consistency check class used in debug mode to check whether log_lsa's are applied in ever increasing order on a per-vpid basis;

log_recovery.h:
- added a new header and moved all functions implemented in the corresponding `.c` file and which need to be exposed for other areas
- adapted includes where needed

log_recovery.c:
- renamed function `log_rv_redo_record_sync_or_dispatch_parallel` to `log_rv_redo_record_sync_or_dispatch_async`; in a future change this function will become the point where the decision is take to either apply a log redo synchronously or dispatch it asynchronously
- out of function `log_rv_redo_record_sync_or_dispatch_async` extracted all code that will be invoked both sync and async in a new function `log_rv_redo_record_sync` which has been moved to the `log_recovery_redo.hpp` header
- moved templated functions that will be called both sequentially (sync) and async to the `log_recovery_redo.hpp` header
