+++ src/transaction/log_recovery.c
@@ -22,9 +22,10 @@
 
 #ident "$Id$"
 
-#include "config.h"
+#include "log_recovery.h"
 
 #include "boot_sr.h"
+#include "config.h"
 #include "error_manager.h"
 #include "locator_sr.h"
 #include "log_2pc.h"
@@ -35,7 +36,7 @@
 #include "log_manager.h"
 #include "log_reader.hpp"
 #include "log_record.hpp"
-#include "log_recovery_redo_log_rec.hpp"
+#include "log_recovery_redo.hpp"
 #include "log_system_tran.hpp"
 #include "log_volids.hpp"
 #include "message_catalog.h"
@@ -44,7 +45,6 @@
 #include "page_buffer.h"
 #include "porting_inline.hpp"
 #include "recovery.h"
-#include "scope_exit.hpp"
 #include "server_type.hpp"
 #include "slotted_page.h"
 #include "system_parameter.h"
@@ -66,32 +66,17 @@ static void log_rv_redo_record (THREAD_ENTRY * thread_p, log_reader & log_pgptr_
 				int (*redofun) (THREAD_ENTRY * thread_p, LOG_RCV *), LOG_RCV * rcv,
 				const LOG_LSA * rcv_lsa_ptr, int undo_length, const char *undo_data,
 				LOG_ZIP & redo_unzip);
-static bool
-log_rv_fix_page_and_check_redo_is_needed (THREAD_ENTRY * thread_p, const VPID & page_vpid, log_rcv & rcv,
-					  LOG_RCVINDEX rcvindex, const log_lsa & rcv_lsa, const LOG_LSA * end_redo_lsa);
 
+static bool log_rv_need_sync_redo (LOG_RCVINDEX rcvindex);
 // *INDENT-OFF*
 template <typename T>
-static int log_rv_get_log_rec_redo_data (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader, const T &log_rec,
-                                         log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-                                         struct log_zip &redo_unzip_support);
-template <typename T>
-static void log_rv_redo_record_debug_logging (const log_lsa &rcv_lsa, LOG_RCVINDEX rcvindex, const vpid &rcv_vpid,
-                                              const log_rcv &rcv);
-template <typename T>
-static void log_rv_redo_record_sync_or_dispatch_parallel (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-                                                          const T &log_rec, const log_lsa &rcv_lsa,
-                                                          const LOG_LSA *end_redo_lsa, LOG_RECTYPE log_rtype,
-                                                          LOG_ZIP &undo_unzip_support, LOG_ZIP &redo_unzip_support);
+static void log_rv_redo_record_sync_or_dispatch_async (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
+						       const T &log_rec, const log_lsa &rcv_lsa,
+						       const LOG_LSA *end_redo_lsa, LOG_RECTYPE log_rtype,
+						       LOG_ZIP &undo_unzip_support, LOG_ZIP &redo_unzip_support);
 // *INDENT-ON*
 
 static bool log_rv_find_checkpoint (THREAD_ENTRY * thread_p, VOLID volid, LOG_LSA * rcv_lsa);
-
-static int log_rv_get_unzip_log_data (THREAD_ENTRY * thread_p, int length, log_reader & log_pgptr_reader,
-				      LOG_ZIP * unzip_ptr, bool &is_zip);
-static int log_rv_get_unzip_and_diff_redo_log_data (THREAD_ENTRY * thread_p, log_reader & log_pgptr_reader,
-						    LOG_RCV * rcv, int undo_length, const char *undo_data,
-						    LOG_ZIP & redo_unzip);
 
 static int log_rv_analysis_undo_redo (THREAD_ENTRY * thread_p, int tran_id, LOG_LSA * log_lsa);
 static int log_rv_analysis_dummy_head_postpone (THREAD_ENTRY * thread_p, int tran_id, LOG_LSA * log_lsa);
@@ -161,6 +146,7 @@ STATIC_INLINE PAGE_PTR log_rv_redo_fix_page (THREAD_ENTRY * thread_p, const VPID
 
 static void log_rv_simulate_runtime_worker (THREAD_ENTRY * thread_p, LOG_TDES * tdes);
 static void log_rv_end_simulation (THREAD_ENTRY * thread_p);
+static void log_find_unilaterally_largest_undo_lsa (THREAD_ENTRY * thread_p, LOG_LSA & max_undo_lsa);
 
 /*
  * CRASH RECOVERY PROCESS
@@ -512,7 +498,17 @@ log_rv_redo_record (THREAD_ENTRY * thread_p, log_reader & log_pgptr_reader,
     }
 }
 
-/* TODO: desc
+/* log_rv_fix_page_and_check_redo_is_needed - check if page still exists and, if yes, whether the lsa has not
+ *                    already been applied
+ *
+ * return: boolean
+ *
+ *  thread_p(in):
+ *  page_vpid(in): page identifier
+ *  rcv(in/out):
+ *  rcvindex(in): recovery index of log record to redo
+ *  rcv_lsa(in): Reset data page (rcv->pgptr) to this LSA
+ *  end_redo_lsa(in):
  */
 bool
 log_rv_fix_page_and_check_redo_is_needed (THREAD_ENTRY * thread_p, const VPID & page_vpid, log_rcv & rcv,
@@ -554,158 +550,7 @@ log_rv_fix_page_and_check_redo_is_needed (THREAD_ENTRY * thread_p, const VPID & 
 
 // *INDENT-OFF*
 /*
- * log_rv_get_log_rec_redo_data - GET UNZIPPED [and DIFFED, if needed] REDO LOG DATA FROM LOG
- *
- * return: error code
- *
- *   thread_p(in):
- *   log_pgptr_reader(in/out): log reader
- *   log_rec(in): log record structure with info about the location and size of the data in the log page
- *   rcv(in/out): Recovery structure for recovery function
- *   log_rtype(in): log record type needed to check if diff information is needed or should be skipped
- *   undo_unzip_support(out): extracted undo data support structure (set as a side effect)
- *   redo_unzip_support(out): extracted redo data support structure (set as a side effect); required to
- *                    be passed by address because it also functions as an internal working buffer
- */
-template <typename T>
-int log_rv_get_log_rec_redo_data (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader, const T &log_rec,
-    log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support, struct log_zip &redo_unzip_support)
-{
-  static_assert (sizeof (T) == 0, "should not be called");
-  return -1;
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_UNDOREDO> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_UNDOREDO &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  /* current log reader position is aligned at the undo data, redo data follows (aligned) the undo data
-   */
-  const bool need_diff_with_undo = (log_rtype == LOG_DIFF_UNDOREDO_DATA || log_rtype == LOG_MVCC_DIFF_UNDOREDO_DATA);
-  if (need_diff_with_undo)
-    {
-      /* for the diff log records, undo and redo data must be read, the diff be applied between the undo and the redo
-       * to reconstruct the actual redo data
-       */
-      bool dummy_is_zip = false;
-      const int err_undo_unzip =
-	log_rv_get_unzip_log_data (thread_p, log_rec.ulength, log_pgptr_reader, &undo_unzip_support,
-				   dummy_is_zip);
-      if (err_undo_unzip != NO_ERROR)
-	{
-	  log_Gl.unique_stats_table.curr_rcv_rec_lsa.set_null ();
-	  logpb_fatal_error (thread_p, true, ARG_FILE_LINE, "log_rv_get_log_rec_redo_data");
-	  return err_undo_unzip;
-	}
-      log_pgptr_reader.align ();
-      return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, undo_unzip_support.data_length,
-						      undo_unzip_support.log_data, redo_unzip_support);
-    }
-  else
-    {
-      /* for the non-diff log records, it is enough to skip undo data and read the redo data
-       */
-      const int temp_length = GET_ZIP_LEN (log_rec.ulength);
-      const int err_skip = log_pgptr_reader.skip (temp_length);
-      if (err_skip != NO_ERROR)
-	{
-	  log_Gl.unique_stats_table.curr_rcv_rec_lsa.set_null ();
-	  logpb_fatal_error (thread_p, true, ARG_FILE_LINE, "log_rv_get_log_rec_redo_data");
-	  return ER_FAILED;
-	}
-      log_pgptr_reader.align ();
-      return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, 0, nullptr, redo_unzip_support);
-    }
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_MVCC_UNDOREDO> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_MVCC_UNDOREDO &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  return log_rv_get_log_rec_redo_data<LOG_REC_UNDOREDO> (thread_p, log_pgptr_reader, log_rec.undoredo,
-                                                         rcv, log_rtype, undo_unzip_support, redo_unzip_support);
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_MVCC_REDO> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_MVCC_REDO &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, 0, nullptr, redo_unzip_support);
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_REDO> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_REDO &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, 0, nullptr, redo_unzip_support);
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_RUN_POSTPONE> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_RUN_POSTPONE &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, 0, nullptr, redo_unzip_support);
-}
-
-template <>
-int log_rv_get_log_rec_redo_data<LOG_REC_COMPENSATE> (THREAD_ENTRY *thread_p, log_reader &log_pgptr_reader,
-    const LOG_REC_COMPENSATE &log_rec, log_rcv &rcv, log_rectype log_rtype, struct log_zip &undo_unzip_support,
-    struct log_zip &redo_unzip_support)
-{
-  return log_rv_get_unzip_and_diff_redo_log_data (thread_p, log_pgptr_reader, &rcv, 0, nullptr, redo_unzip_support);
-}
-// *INDENT-ON*
-
-#if !defined(NDEBUG)
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_MVCC_UNDOREDO);
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_UNDOREDO);
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_MVCC_REDO);
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_REDO);
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_RUN_POSTPONE);
-DBG_REGISTER_PARSE_TYPE_NAME (LOG_REC_COMPENSATE);
-#endif
-
-// *INDENT-OFF*
-/* log_rv_redo_record_debug_logging - utility function which prints debug information
- *                        templated in order to be able to specifically print out user readable name
- *                        of the log record structure
- */
-template <typename T>
-void log_rv_redo_record_debug_logging (const log_lsa & rcv_lsa, LOG_RCVINDEX rcvindex, const vpid & rcv_vpid,
-                                       const log_rcv & rcv)
-{
-#if !defined(NDEBUG)
-  if (prm_get_bool_value (PRM_ID_LOG_TRACE_DEBUG))
-    {
-      constexpr const char *const log_rec_name = dbg_parse_type_name<T> ();
-      fprintf (stdout,
-	       "TRACE REDOING[%s]: LSA = %lld|%d, Rv_index = %s,\n"
-	       "      volid = %d, pageid = %d, offset = %d,\n", log_rec_name, LSA_AS_ARGS (&rcv_lsa),
-	       rv_rcvindex_string (rcvindex), rcv_vpid.volid, rcv_vpid.pageid, rcv.offset);
-      if (rcv.pgptr != nullptr)
-	{
-	  const log_lsa *const rcv_page_lsaptr = pgbuf_get_lsa (rcv.pgptr);
-	  assert (rcv_page_lsaptr != nullptr);
-	  fprintf (stdout, "      page_lsa = %lld|%d\n", LSA_AS_ARGS (rcv_page_lsaptr));
-	}
-      else
-	{
-	  fprintf (stdout, "      page_lsa = %lld|%d\n", -1LL, -1);
-	}
-      fflush (stdout);
-    }
-#endif /* !NDEBUG */
-}
-// *INDENT-ON*
-
-// *INDENT-OFF*
-/*
- * log_rv_redo_record_sync_or_dispatch_parallel - execute a redo record synchronously or
+ * log_rv_redo_record_sync_or_dispatch_async - execute a redo record synchronously or
  *                    (in future) asynchronously
  *
  * return: nothing
@@ -724,83 +569,21 @@ void log_rv_redo_record_debug_logging (const log_lsa & rcv_lsa, LOG_RCVINDEX rcv
  *
  */
 template <typename T>
-void log_rv_redo_record_sync_or_dispatch_parallel (THREAD_ENTRY * thread_p, log_reader & log_pgptr_reader,
-                                                   const T & log_rec, const log_lsa & rcv_lsa,
-                                                   const LOG_LSA * end_redo_lsa, LOG_RECTYPE log_rtype,
-                                                   LOG_ZIP & undo_unzip_support, LOG_ZIP & redo_unzip_support)
+void log_rv_redo_record_sync_or_dispatch_async (THREAD_ENTRY * thread_p, log_reader & log_pgptr_reader,
+                                                const T & log_rec, const log_lsa & rcv_lsa,
+                                                const LOG_LSA * end_redo_lsa, LOG_RECTYPE log_rtype,
+                                                LOG_ZIP & undo_unzip_support, LOG_ZIP & redo_unzip_support)
 {
   const VPID rcv_vpid = log_rv_get_log_rec_vpid<T> (log_rec);
   // at this point, vpid can either be valid or not
 
-  // TODO: once valid vpid is extracted, it can be decided whether to execute sync or dispatch asynchronously
-  // as such, from here on, everything should be moved into another function
+  const LOG_DATA &log_data = log_rv_get_log_rec_data<T> (log_rec);
 
-  const LOG_DATA & log_data = log_rv_get_log_rec_data<T> (log_rec);
+  // once vpid is extracted (or not), and depending on parameters, either dispatch the applying of
+  // log redo asynchronously, or invoke synchronously
 
-  LOG_RCV rcv;
-  if (!log_rv_fix_page_and_check_redo_is_needed (thread_p, rcv_vpid, rcv, log_data.rcvindex, rcv_lsa, end_redo_lsa))
-    {
-      /* nothing else needs to be done, see explanation in function */
-      assert (rcv.pgptr == nullptr);
-      return;
-    }
-  // at this point, pgptr can be null or not
-
-  /* will take care of unfixing the page, will be correctly de-allocated as it is the same
-   * storage class as 'rcv' and allocated on the stack after 'rcv' */
-  scope_exit <std::function<void (void)>> unfix_rcv_pgptr (
-    // could have used pgbuf_unfix_and_init if it were a function
-    [&thread_p, &rcv] ()
-    {
-      if (rcv.pgptr != nullptr)
-        {
-          pgbuf_unfix (thread_p, rcv.pgptr);
-          rcv.pgptr = nullptr;
-        }
-    }
-  );
-
-  rcv.length = log_rv_get_log_rec_redo_length<T> (log_rec);
-  rcv.mvcc_id = log_rv_get_log_rec_mvccid<T> (log_rec);
-  rcv.offset = log_rv_get_log_rec_offset<T> (log_rec);
-
-  log_rv_redo_record_debug_logging<T> (rcv_lsa, log_data.rcvindex, rcv_vpid, rcv);
-
-  const auto err_redo_data = log_rv_get_log_rec_redo_data<T> (thread_p, log_pgptr_reader, log_rec, rcv, log_rtype,
-                                                              undo_unzip_support, redo_unzip_support);
-  if (err_redo_data != NO_ERROR)
-    {
-      logpb_fatal_error (thread_p, true, ARG_FILE_LINE, "log_rv_get_log_rec_redo_data");
-      // rcv pgptr will be automatically unfixed
-      return;
-    }
-
-  rvfun::fun_t redofunc = log_rv_get_fun<T> (log_rec, log_data.rcvindex);
-  if (redofunc != nullptr)
-    {
-      const int err_func = redofunc (thread_p, &rcv);
-      if (err_func != NO_ERROR)
-	{
-	  logpb_fatal_error (thread_p, true, ARG_FILE_LINE,
-			     "log_rv_redo_record_sync_or_dispatch_parallel: Error applying redo record at log_lsa=(%lld, %d), "
-			     "rcv = {mvccid=%llu, vpid=(%d, %d), offset = %d, data_length = %d}",
-			     (long long int) rcv_lsa.pageid, (int) rcv_lsa.offset,
-			     (long long int) rcv.mvcc_id, (int) rcv_vpid.pageid, (int) rcv_vpid.volid,
-			     (int) rcv.offset, (int) rcv.length);
-	}
-    }
-  else
-    {
-      er_log_debug (ARG_FILE_LINE,
-		    "log_rv_redo_record_sync_or_dispatch_parallel: WARNING.. There is not a"
-		    " REDO (or, possibly, UNDO) function to execute. May produce recovery problems.");
-    }
-
-  if (rcv.pgptr != nullptr)
-    {
-      pgbuf_set_lsa (thread_p, rcv.pgptr, &rcv_lsa);
-      // rcv pgptr will be automatically unfixed
-    }
+  log_rv_redo_record_sync<T>(thread_p, log_pgptr_reader, log_rec, rcv_vpid, rcv_lsa, end_redo_lsa, log_rtype,
+                             undo_unzip_support, redo_unzip_support);
 }
 // *INDENT-ON*
 
@@ -845,7 +628,7 @@ log_rv_find_checkpoint (THREAD_ENTRY * thread_p, VOLID volid, LOG_LSA * rcv_lsa)
  * NOTE:if log_data is unzip data return LOG_ZIP data
  *               else log_data is zip data return unzip log data
  */
-static int
+int
 log_rv_get_unzip_log_data (THREAD_ENTRY * thread_p, int length, log_reader & log_pgptr_reader,
 			   LOG_ZIP * unzip_ptr, bool &is_zip)
 {
@@ -935,7 +718,7 @@ log_rv_get_unzip_log_data (THREAD_ENTRY * thread_p, int length, log_reader & log
  * TODO: after refactoring, it's easier to supply (or not) a pointer to the undo unzip
  * TODO: separate in 2 based on diffing or not and fork outside
  */
-static int
+int
 log_rv_get_unzip_and_diff_redo_log_data (THREAD_ENTRY * thread_p, log_reader & log_pgptr_reader,
 					 LOG_RCV * rcv, int undo_length, const char *undo_data, LOG_ZIP & redo_unzip)
 {
@@ -3478,20 +3261,20 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 
 	  LSA_COPY (&log_Gl.unique_stats_table.curr_rcv_rec_lsa, &lsa);
 
-	  /* Find the log record */
+	  /* both the page id and the offsset might have changed; page id is changed at the end of the loop */
 	  log_pgptr_reader.set_lsa_and_fetch_page (lsa);
 
 	  {
 	    /* Pointer to log record */
 	    // *INDENT-OFF*
-	    const LOG_RECORD_HEADER *log_rec = log_pgptr_reader.reinterpret_cptr<LOG_RECORD_HEADER> ();
+	    const LOG_RECORD_HEADER *log_rec_header = log_pgptr_reader.reinterpret_cptr<LOG_RECORD_HEADER> ();
 	    // *INDENT-ON*
 
-	    tran_id = log_rec->trid;
-	    log_rtype = log_rec->type;
+	    tran_id = log_rec_header->trid;
+	    log_rtype = log_rec_header->type;
 
 	    /* Get the address of next log record to scan */
-	    LSA_COPY (&lsa, &log_rec->forw_lsa);
+	    LSA_COPY (&lsa, &log_rec_header->forw_lsa);
 	  }
 
 	  /*
@@ -3557,11 +3340,11 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		LSA_COPY (&log_Gl.hdr.mvcc_op_log_lsa, &rcv_lsa);
 
                 // *INDENT-OFF*
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_MVCC_UNDOREDO> (thread_p, log_pgptr_reader,
-                                                                                     log_rec_mvcc_undoredo,
-                                                                                     rcv_lsa, end_redo_lsa,
-                                                                                     log_rtype, *undo_unzip_ptr,
-                                                                                     *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_MVCC_UNDOREDO> (thread_p, log_pgptr_reader,
+                                                                                  log_rec_mvcc_undoredo,
+                                                                                  rcv_lsa, end_redo_lsa,
+                                                                                  log_rtype, *undo_unzip_ptr,
+                                                                                  *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -3580,10 +3363,10 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		const LOG_REC_UNDOREDO log_rec_undoredo
 		    = log_pgptr_reader.reinterpret_copy_and_add_align<LOG_REC_UNDOREDO>();
 
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_UNDOREDO> (thread_p, log_pgptr_reader,
-                                                                                log_rec_undoredo, rcv_lsa,
-                                                                                end_redo_lsa, log_rtype,
-                                                                                *undo_unzip_ptr, *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_UNDOREDO> (thread_p, log_pgptr_reader,
+                                                                             log_rec_undoredo, rcv_lsa,
+                                                                             end_redo_lsa, log_rtype,
+                                                                             *undo_unzip_ptr, *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -3615,11 +3398,10 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		 * relevant for vacuum as vacuum only processes undo data */
 
                 // *INDENT-OFF*
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_MVCC_REDO> (thread_p, log_pgptr_reader,
-                                                                                 log_rec_mvcc_redo,
-                                                                                 rcv_lsa, end_redo_lsa,
-                                                                                 log_rtype, *undo_unzip_ptr,
-                                                                                 *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_MVCC_REDO> (thread_p, log_pgptr_reader,
+                                                                              log_rec_mvcc_redo, rcv_lsa, end_redo_lsa,
+                                                                              log_rtype, *undo_unzip_ptr,
+                                                                              *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -3642,11 +3424,10 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		    logpb_vacuum_reset_log_header_cache (thread_p, &log_Gl.hdr);
 		  }
 
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_REDO> (thread_p, log_pgptr_reader,
-                                                                            log_rec_redo,
-                                                                            rcv_lsa, end_redo_lsa,
-                                                                            log_rtype, *undo_unzip_ptr,
-                                                                            *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_REDO> (thread_p, log_pgptr_reader,
+                                                                         log_rec_redo, rcv_lsa, end_redo_lsa,
+                                                                         log_rtype, *undo_unzip_ptr,
+                                                                         *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -3705,11 +3486,10 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		const LOG_REC_RUN_POSTPONE log_rec_run_posp
 		    = log_pgptr_reader.reinterpret_copy_and_add_align<LOG_REC_RUN_POSTPONE>();
 
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_RUN_POSTPONE> (thread_p, log_pgptr_reader,
-                                                                                    log_rec_run_posp,
-                                                                                    rcv_lsa, end_redo_lsa,
-                                                                                    log_rtype, *undo_unzip_ptr,
-                                                                                    *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_RUN_POSTPONE> (thread_p, log_pgptr_reader,
+                                                                                 log_rec_run_posp, rcv_lsa,
+                                                                                 end_redo_lsa, log_rtype,
+                                                                                 *undo_unzip_ptr, *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -3726,11 +3506,10 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 		const LOG_REC_COMPENSATE log_rec_compensate
 		    = log_pgptr_reader.reinterpret_copy_and_add_align<LOG_REC_COMPENSATE>();
 
-                log_rv_redo_record_sync_or_dispatch_parallel<LOG_REC_COMPENSATE> (thread_p, log_pgptr_reader,
-                                                                                  log_rec_compensate,
-                                                                                  rcv_lsa, end_redo_lsa,
-                                                                                  log_rtype, *undo_unzip_ptr,
-                                                                                  *redo_unzip_ptr);
+                log_rv_redo_record_sync_or_dispatch_async<LOG_REC_COMPENSATE> (thread_p, log_pgptr_reader,
+                                                                               log_rec_compensate, rcv_lsa,
+                                                                               end_redo_lsa, log_rtype,
+                                                                               *undo_unzip_ptr, *redo_unzip_ptr);
                 // *INDENT-ON*
 	      }
 	      break;
@@ -4055,6 +3834,12 @@ log_recovery_redo (THREAD_ENTRY * thread_p, const LOG_LSA * start_redolsa, const
 
 exit:
   LSA_SET_NULL (&log_Gl.unique_stats_table.curr_rcv_rec_lsa);
+
+#if !defined(NDEBUG)
+  // bit of debug code to ensure that, should this code be executed asynchronously, within the same page,
+  // the lsa is ever-increasing, thus, not altering the order in which it has been added to the log in the first place
+  log_Gl_recovery_redo_consistency_check.cleanup ();
+#endif
 
   return;
 }
@@ -6194,7 +5979,7 @@ log_rv_undo_record_modify (THREAD_ENTRY * thread_p, LOG_RCV * rcv)
 }
 
 /*
- * log_rv_redo_record_modify () - Modify one record of database slotted page.
+ * log_rv_record_modify_internal () - Modify one record of database slotted page.
  *				  The change can be one of:
  *				  1. New record is inserted.
  *				  2. Existing record is removed.
@@ -6338,7 +6123,7 @@ log_rv_pack_redo_record_changes (char *ptr, int offset_to_data, int old_data_siz
 }
 
 /*
- * log_rv_pack_redo_record_changes () - Pack recovery data for undo record
+ * log_rv_pack_undo_record_changes () - Pack recovery data for undo record
  *					change.
  *
  * return	       : Error code.
@@ -6498,3 +6283,51 @@ log_rv_end_simulation (THREAD_ENTRY * thread_p)
   LOG_SET_CURRENT_TRAN_INDEX (thread_p, LOG_SYSTEM_TRAN_INDEX);
 #endif // SA_MODE
 }
+
+/*
+ * log_find_unilaterally_largest_undo_lsa - find maximum lsa address to undo
+ *
+ * return:
+ *
+ * Note: Find the maximum log sequence address to undo during the undo
+ *              crash recovery phase.
+ */
+void
+log_find_unilaterally_largest_undo_lsa (THREAD_ENTRY * thread_p, LOG_LSA & max_undo_lsa)
+{
+  // *INDENT-OFF*
+  int i;
+  LOG_TDES *tdes;		/* Transaction descriptor */
+
+  TR_TABLE_CS_ENTER_READ_MODE (thread_p);
+
+  LSA_SET_NULL (&max_undo_lsa);
+
+  auto max_undo_lsa_func = [&] (log_tdes & tdes)
+    {
+      if (LSA_LT (&max_undo_lsa, &tdes.undo_nxlsa))
+        {
+          max_undo_lsa = tdes.undo_nxlsa;
+        }
+    };
+
+  /* Check active transactions. */
+  for (i = 0; i < log_Gl.trantable.num_total_indices; i++)
+    {
+      if (i != LOG_SYSTEM_TRAN_INDEX)
+        {
+          tdes = log_Gl.trantable.all_tdes[i];
+          if (tdes != NULL && tdes->trid != NULL_TRANID
+              && (tdes->state == TRAN_UNACTIVE_UNILATERALLY_ABORTED || tdes->state == TRAN_UNACTIVE_ABORTED))
+            {
+              max_undo_lsa_func (*tdes);
+            }
+        }
+    }
+  /* Check system worker transactions. */
+  log_system_tdes::map_all_tdes (max_undo_lsa_func);
+
+  TR_TABLE_CS_EXIT (thread_p);
+  // *INDENT-ON*
+}
+
