+++ src/transaction/log_tran_table.c
@@ -4161,53 +4161,6 @@ logtb_set_num_loose_end_trans (THREAD_ENTRY * thread_p)
 }
 
 /*
- * log_find_unilaterally_largest_undo_lsa - find maximum lsa address to undo
- *
- * return:
- *
- * Note: Find the maximum log sequence address to undo during the undo
- *              crash recovery phase.
- */
-void
-log_find_unilaterally_largest_undo_lsa (THREAD_ENTRY * thread_p, LOG_LSA & max_undo_lsa)
-{
-  // *INDENT-OFF*
-  int i;
-  LOG_TDES *tdes;		/* Transaction descriptor */
-
-  TR_TABLE_CS_ENTER_READ_MODE (thread_p);
-
-  LSA_SET_NULL (&max_undo_lsa);
-
-  auto max_undo_lsa_func = [&] (log_tdes & tdes)
-    {
-      if (LSA_LT (&max_undo_lsa, &tdes.undo_nxlsa))
-        {
-          max_undo_lsa = tdes.undo_nxlsa;
-        }
-    };
-
-  /* Check active transactions. */
-  for (i = 0; i < NUM_TOTAL_TRAN_INDICES; i++)
-    {
-      if (i != LOG_SYSTEM_TRAN_INDEX)
-	{
-	  tdes = log_Gl.trantable.all_tdes[i];
-	  if (tdes != NULL && tdes->trid != NULL_TRANID
-	      && (tdes->state == TRAN_UNACTIVE_UNILATERALLY_ABORTED || tdes->state == TRAN_UNACTIVE_ABORTED))
-	    {
-	      max_undo_lsa_func (*tdes);
-	    }
-	}
-    }
-  /* Check system worker transactions. */
-  log_system_tdes::map_all_tdes (max_undo_lsa_func);
-
-  TR_TABLE_CS_EXIT (thread_p);
-  // *INDENT-ON*
-}
-
-/*
  * logtb_find_smallest_lsa - smallest lsa address of all active transactions
  *
  * return:
